Move creation of reader and writer out of TransactionEngine
THe creation of the reader and writer is now done in main.rs, so that
TransactionEngine does not have to be concerned with which file to read
from or output to write to.
Removed lib.rs and moved the code to main.rs to consolidate.
Round the Decimal values of an account at serialization time, instead of
doing it preventively before writing to output.